groups: the pose table is a storage buffer — whole-row partial uploads
The per-group table (pose, color, scale, clip + the far-slab lanes) was two
RGBA32F DataTextures — a WebGL2-era idiom with two structural costs: the
WebGPU backend ignores texture update ranges, so ANY dirty group re-uploaded
the full 1.28MB table that frame; and the texture dimension limit walled the
table at 16k rows. Now it is ONE storage buffer: GROUP_STRIDE(7) vec4 rows
(5 pose cols + the 2 far cols fused), schema and constants in
core/glyphVertex.js beside the primary reader. Every consumer reads it as
indexed vec4 elements — the shared glyph transform (render, occluder, pick),
GlyphField's far-lane varyings, and PanelField's render + pick materials.
Setters mark whole dirty rows via _touchGroup -> addUpdateRange, so a moved
group uploads ~112 BYTES. Growth doubles (cap 2^20 rows, ~117MB backstop —
the 16k wall is gone) and rides a new registry seam: every group-reading
material registers (registerGroupMaterial), growth swaps the attribute and
disposes them all (three's bind-group cache is texture-keyed; a swapped
storage buffer under a live bind group otherwise submits the old buffer
forever). PanelField unregisters its per-instance materials on dispose.
Robust storage access CLAMPS out-of-range indices where textureLoad returned
zeros, so both vertex transforms grew an explicit bound cull (a stale id past
capacity must not ghost-render wearing the last row's pose).

Two live-verified bugs fell out of the bring-up, one of them ancient:

- _touchGroup's version bump is UNCONDITIONAL. Buffer creation uploads the
  whole array but leaves marked ranges behind; a version gated on dirty-set
  membership deadlocked - rows marked before first bind blocked their own
  future updates forever (poses frozen mid-glide: the tiny-grids screenshot).

- The pose sweep's matrixWorld cache was Float32Array. Matrix elements are
  f64; the f32 cache truncated on set and the exact compare then failed
  FOREVER for any non-f32-representable value, re-posing every group in every
  sweep (x3 meshes) every frame. On the texture path this had silently meant
  a full-table re-upload every frame since the sweep was born - the buffer's
  dirty tracking made it visible. The cache is Float64Array now.

Measured on a 1,018-file pour (4.6k groups): idle group traffic 0 bytes/frame
(was 1.28MB/frame with the broken gate), one moved row = its 112B range,
poses match matrixWorld for all 3,558 views, zero page errors. Next-biggest
per-frame upload is now the range-culler's indirect record rewrite (~668KB/f)
- a follow-on, not this change. Gates: group-reuse grew partial-upload + f64
+ no-deadlock teeth (29); all suites + loadstorm green; live pour + camera
fitall pixel-verified.

+++ packages/glyph3d-core/src/MegaGlyphField.js
@@ -340,7 +340,10 @@ export class MegaGlyphField {
      * @returns {{groupId: number, release: () => void}}
      */
     createPoseGroup(node) {
-        const entry = { node, groupId: this.field.createGroup(), _mat: new Float32Array(16).fill(NaN), dead: false };
+        // Float64 cache — matrixWorld.elements are f64; an f32 cache truncates on
+        // set and the exact compare then fails FOREVER on any non-f32 value,
+        // re-posing every group every sweep (the every-frame full-upload bug).
+        const entry = { node, groupId: this.field.createGroup(), _mat: new Float64Array(16).fill(NaN), dead: false };
         this._poseGroups.push(entry);
         const mega = this;
         return {
@@ -536,7 +539,10 @@ export class MegaFieldView {
         this._visible = true;
         this._nodeVisible = true;   // scene-graph visibility, mirrored by the pose sweep
         this._alpha = 1;
-        this._mat = new Float32Array(16).fill(NaN); // NaN ≠ anything → first sweep always poses
+        // Float64: matrixWorld.elements are f64 — an f32 cache truncates on set and
+        // the exact compare fails forever (every-frame reposing). NaN ≠ anything →
+        // the first sweep always poses.
+        this._mat = new Float64Array(16).fill(NaN);
     }
 
     /** The arena's attach seam — stage()/adoptField and realloc re-attach call this.
